online-banking-flask-app/requirements.txt
Flask==2.3.2
bcrypt==4.0.1
PyJWT==2.8.0
